scripts/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0

langchain==0.1.1
langchain-community>=0.0.13,<0.1
langchain-core==0.1.23
langsmith==0.0.87
langgraph==0.0.24

httpx==0.25.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
loguru==0.7.2
